third_party/blink/web_tests/http/tests/security/mixedContent/insecure-sync-post-xhr-allowed-expected.txt

ALERT: Executing script in the child window.
CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
ALERT: XHR POST opened.
CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-sync-xhr-post.html' was loaded over HTTPS, but requested an insecure XMLHttpRequest endpoint 'http://example.test:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*&headers=Content-Type'. This content should also be served over HTTPS.
ALERT: XHR POST sent.
This test opens a HTTPS window that loads insecure data via XHR. We should trigger a mixed content callback because the main frame in the window is HTTPS but now has insecure data.
